requirements.txt
flask
gunicorn
mysql-connector-python
pandas
plotly
numpy
dash
openpyxl